refactor(core): extract profile SDK layers
