+++ scripts/readonly-agent-bootstrap.test.mjs
@@ -266,6 +266,9 @@ test('declared discovery routes are zero-write in a cold read-only fixture', (t)
     'framework/core/src/python/kungfu/distribution_update.py',
     'framework/core/src/python/kungfu/distribution_update_planning.py',
     'framework/core/src/python/kungfu/distribution_update_policy.py',
+    'framework/core/src/python/kungfu/profile_sdk_kfd3.py',
+    'framework/core/src/python/kungfu/profile_sdk_source.py',
+    'framework/core/src/python/kungfu/profile_sdk_support.py',
     'framework/core/src/python/kungfu/exit_bundle.py',
     'framework/core/src/python/kungfu/exit_verifier.py',
     'framework/core/src/python/kungfu/release_channel.py',
